fix(pair): frame is already authorized
