+++ contracts/pair/src/token/mod.rs
@@ -7,12 +7,10 @@ mod contract;
 mod event;
 mod metadata;
 mod storage_types;
-mod internal_mint;
-mod internal_burn;
+mod internal_fn;
 //mod test;
 
 pub use crate::token::contract::TokenClient;
 pub use crate::token::contract::Token;
 pub use crate::token::contract::TokenTrait;
-pub use crate::token::internal_mint::internal_mint;
-pub use crate::token::internal_burn::internal_burn;
+pub use crate::token::internal_fn::{internal_mint, internal_burn, internal_transfer};
